+++ models/src/main/kotlin/io/provenance/api/models/cee/ContractConfig.kt
@@ -1,4 +1,4 @@
-package io.provenance.cee.api.models.cee
+package io.provenance.api.models.cee
 
 import java.util.UUID
