+++ models/src/main/kotlin/io/provenance/api/models/cee/ExecuteContractConfig.kt
@@ -1,0 +1,12 @@
+package io.provenance.api.models.cee
+
+import io.provenance.api.models.eos.ObjectStoreConfig
+import io.provenance.api.models.account.AccountInfo
+import io.provenance.api.models.p8e.ProvenanceConfig
+
+data class ExecuteContractConfig(
+    val contract: ContractConfig,
+    val client: ObjectStoreConfig,
+    val account: AccountInfo,
+    val provenanceConfig: ProvenanceConfig,
+)
