+++ models/src/main/kotlin/io/provenance/api/models/cee/ExecuteContractRequest.kt
@@ -1,6 +1,6 @@
-package io.provenance.cee.api.models.cee
+package io.provenance.api.models.cee
 
-import io.provenance.cee.api.models.account.AccountInfo
+import io.provenance.api.models.account.AccountInfo
 
 data class ExecuteContractRequest(
     val config: ExecuteContractConfig,
