+++ models/src/main/kotlin/io/provenance/api/models/eos/ObjectStoreConfig.kt
@@ -1,4 +1,4 @@
-package io.provenance.cee.api.models.eos
+package io.provenance.api.models.eos
 
 data class ObjectStoreConfig(
     val objectStoreUrl: String,
