+++ service/src/main/kotlin/io/provenance/onboarding/domain/usecase/cee/common/client/model/CreateClientRequest.kt
@@ -1,7 +1,7 @@
 package io.provenance.onboarding.domain.usecase.cee.common.client.model
 
-import io.provenance.cee.api.models.account.AccountInfo
-import io.provenance.cee.api.models.eos.ObjectStoreConfig
+import io.provenance.api.models.account.AccountInfo
+import io.provenance.api.models.eos.ObjectStoreConfig
 
 data class CreateClientRequest(
     val account: AccountInfo,
