+++ service/src/main/kotlin/io/provenance/onboarding/domain/usecase/objectStore/model/SnapshotAssetRequest.kt
@@ -1,6 +1,6 @@
 package io.provenance.onboarding.domain.usecase.objectStore.model
 
-import io.provenance.onboarding.domain.usecase.common.model.AccountInfo
+import io.provenance.api.models.account.AccountInfo
 import io.provenance.onboarding.domain.usecase.common.model.PermissionInfo
 
 data class SnapshotAssetRequest(
