+++ service/src/main/kotlin/io/provenance/onboarding/domain/usecase/provenance/tx/ExecuteTx.kt
@@ -1,8 +1,8 @@
 package io.provenance.onboarding.domain.usecase.provenance.tx
 
+import io.provenance.api.models.p8e.TxResponse
 import io.provenance.onboarding.domain.provenance.Provenance
 import io.provenance.onboarding.domain.usecase.AbstractUseCase
-import io.provenance.onboarding.domain.usecase.common.model.TxResponse
 import io.provenance.onboarding.domain.usecase.provenance.account.GetAccount
 import io.provenance.onboarding.domain.usecase.provenance.tx.model.ExecuteTxRequest
 import org.springframework.stereotype.Component
