+++ service/src/main/kotlin/io/provenance/onboarding/frameworks/provenance/ProvenanceService.kt
@@ -5,6 +5,8 @@ import com.google.protobuf.Any
 import cosmos.base.abci.v1beta1.Abci
 import cosmos.tx.v1beta1.ServiceOuterClass
 import cosmos.tx.v1beta1.TxOuterClass
+import io.provenance.api.models.p8e.ProvenanceConfig
+import io.provenance.api.models.p8e.TxResponse
 import io.provenance.client.grpc.BaseReqSigner
 import io.provenance.client.grpc.GasEstimationMethod
 import io.provenance.client.grpc.PbClient
@@ -13,9 +15,7 @@ import io.provenance.hdwallet.wallet.Account
 import io.provenance.metadata.v1.ScopeRequest
 import io.provenance.metadata.v1.ScopeResponse
 import io.provenance.onboarding.domain.provenance.Provenance
-import io.provenance.onboarding.domain.usecase.common.model.ProvenanceConfig
 import io.provenance.onboarding.domain.usecase.common.model.TxBody
-import io.provenance.onboarding.domain.usecase.common.model.TxResponse
 import io.provenance.onboarding.frameworks.provenance.exceptions.ContractTxException
 import io.provenance.onboarding.frameworks.provenance.extensions.getBaseAccount
 import io.provenance.onboarding.frameworks.provenance.extensions.getCurrentHeight
